+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix issue with corn going from stage 2 -> 3
 - add aloe
 - add blueberry
+- add chrysanthemum
 
 ## [0.2.0] - 2022-05-01