+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add aloe
 - add blueberry
 - add chrysanthemum
+- add coffee
 
 ## [0.2.0] - 2022-05-01