+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add chrysanthemum
 - add coffee
 - add cotton
+- add goldenrod
 
 ## [0.2.0] - 2022-05-01