+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add cotton
 - add goldenrod
 - add gracecorn
+- add hop
 
 ## [0.2.0] - 2022-05-01