+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add goldenrod
 - add gracecorn
 - add hop
+- add potato
 
 ## [0.2.0] - 2022-05-01