+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - ?
+
+- adjust localization for lvl 3 Living off the Land
+- TODO: add craftable schematics to learn traits
+
 ## [0.4.1] - 2022-05-06
 
 - remove unused file in CrystalHell folder